fixed autgh and gitrhub token copy operations within installer

+++ installer.lua
@@ -45,8 +45,12 @@ end
 
 print("\nSuccess. Checking for auth file on disk...")
 if fs.exists("/disk/.auth") then
-    fs.copy("/disk/.auth", "/.auth")
-    print("Authorisation file copied to root. Can be altered at /.auth\n")
+    if fs.exists("/.auth") then
+        print("Auythorisation file found in root. Skipping copy.")
+    else
+        fs.copy("/disk/.auth", "/.auth")
+        print("Authorisation file copied to root. Can be altered at /.auth\n")
+    end
 else
     print(
         "Warning: Could not find .auth file on disk. Please write a custom one to your root directory containing a single line of characters representing your authorisation key that must be consistent between all nodes in your network.\n")
@@ -54,8 +58,12 @@ end
 
 print("Checking for GitHub token file on disk...")
 if fs.exists("/disk/.github") then
-    fs.copy("/disk/.github", "/.github")
-    print("GitHub token file copied to root. Can be altered at /.github\n")
+    if fs.exist("/.github") then
+        print("Github token found in root. Skipping copy.")
+    else
+        fs.copy("/disk/.github", "/.github")
+        print("GitHub token file copied to root. Can be altered at /.github\n")
+    end
 else
     print(
         "Warning: Could not find .github file on disk. Please write a custom one to your root directory containing a single line of youre discord personal access token. This is needed to bypass ratelimits.\n")
